Suggested OpenMP 5.0 implementation
Allocate memory on the host, and map if the device exists.
Rely on the compiler to ignore this is there is no device.
Then use target on all kernels, again hoping the compiler ignores
when no device.
Finally, use the teams distribute parallel for simd on all kernel
loops so that the compiler can "do the right thing".